+++ hurocon/cli/impl/__init__.py
@@ -1,4 +1,0 @@
-"""
-This module contains implementation for some API interactions, that requires
-additional processing before streaming to output.
-"""