+++ hurocon/cli/lte.py
@@ -1,12 +1,16 @@
+"""
+This module is deprecated and will be removed in version `1.0.0`.
+Use `net cellular` cli group instead.
+"""
+
 import click
 from click_didyoumean import DYMGroup
-from huawei_lte_api.Client import Client
 
-from .__cli_base__ import cli
-from ..core.connection import HRC_Connection
+from .cli_base import cli
+from ..implementation import net_impl
 
 
-@cli.group(cls=DYMGroup)
+@cli.group(cls=DYMGroup, hidden=True)
 def lte():
     """
     [DEPRECATED] Cellular connection controls
@@ -19,16 +23,7 @@ def lte():
 @lte.command('status')
 def lte_status():
     """ Get cellular connection status """
-    try:
-        with HRC_Connection() as conn:
-            con_stat = Client(conn).dial_up.mobile_dataswitch()['dataswitch']
-    except Exception as e:
-        msg = 'Execution failed, reason: "{}"'.format(e)
-    else:
-        msg = 'Connected to cellular network' if con_stat == '1' else \
-              'No connection to cellular network'
-
-    click.echo(msg)
+    net_impl.cellular_status_impl()
 
 
 @lte.command('set')
@@ -39,12 +34,4 @@ def lte_set_connection(mode: bool):
 
     MODE (bool): True, False | [Y]es, [N]o | 1, 0
     """
-    try:
-        with HRC_Connection() as conn:
-            Client(conn).dial_up.set_mobile_dataswitch(int(mode))
-    except Exception as e:
-        msg = 'Can not switch connection mode, reason: "{}"'.format(e)
-    else:
-        msg = 'Successfully {} cellular data'.format('enabled' if mode else 'disabled')
-
-    click.echo(msg)
+    net_impl.cellular_set_connection_impl(mode)